requirements.txt
aniso8601==9.0.1
cffi==1.14.5
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
PyNaCl==1.4.0
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.18
Werkzeug==2.0.1
WTForms==2.3.3
